# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorResponseCode = typeof ErrorResponseCode[keyof typeof ErrorResponseCode];


export const ErrorResponseCode = {
  bad_request: 'bad_request',
  quota: 'quota',
  busy: 'busy',
  invalid_output: 'invalid_output',
  blocked: 'blocked',
  unavailable: 'unavailable',
} as const;

export interface ErrorResponse {
  /** Calm, plain-language message that can be shown to the user. */
  error: string;
  code: ErrorResponseCode;
}

export interface ProcessFrame {
  /** @minLength 32 */
  dataUrl: string;
  /** @minimum 0 */
  timestamp: number;
  /** @minimum 0 */
  sharpness: number;
  /** @minimum 0 */
  difference: number;
}

export interface ProcessDocumentInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  documentName: string;
  /**
     * @minItems 1
     * @maxItems 30
     */
  frames: ProcessFrame[];
}

/**
 * Region of an image as [ymin, xmin, ymax, xmax], each 0-1000 (fractions of the image height/width times 1000).
 * @minItems 4
 * @maxItems 4
 * @items.minimum 0
 * @items.maximum 1000
 */
export type Box = number[];

export type PageBlockType = typeof PageBlockType[keyof typeof PageBlockType];


export const PageBlockType = {
  heading: 'heading',
  paragraph: 'paragraph',
  list: 'list',
  table: 'table',
  lines: 'lines',
  caption: 'caption',
  figure: 'figure',
  header: 'header',
  footer: 'footer',
} as const;

/**
 * One piece of a page in reading order, so the page's structure can be rebuilt.
 */
export interface PageBlock {
  type: PageBlockType;
  /**
     * Heading level (1 is the largest). Only for headings.
     * @minimum 1
     * @maximum 3
     */
  level?: number;
  /** Text of the block. For type "lines" each line of the page is on its own line. */
  text: string;
  /** List items without their bullet or number. Only for lists. */
  items?: string[];
  /** Table rows, each a list of cell texts. Only for tables. */
  rows?: string[][];
  /** Where the block is in the best frame, or null if unknown. */
  box: Box | null;
}

export interface ReconstructedPage {
  /** @minimum 1 */
  pageNumber: number;
  title: string;
  text: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  needsReview: boolean;
  /** @nullable */
  reviewReason: string | null;
  /**
     * Zero-based indices (into the request frames) of every frame that shows this page.
     * @minItems 1
     * @items.minimum 0
     */
  sourceFrameIndices: number[];
  /**
     * Zero-based index of the frame that shows this page most clearly. Always one of sourceFrameIndices.
     * @minimum 0
     */
  bestFrameIndex: number;
  /** Where the paper page is inside the best frame, or null if unknown. */
  pageBox: Box | null;
  /** The page's content in reading order. The same text as `text`, with structure and position. */
  blocks: PageBlock[];
  /** Gemini model that produced this page. */
  modelUsed?: string;
}

/**
 * What this request used, for working out running costs.
 */
export interface Usage {
  model: string;
  inputTokens: number;
  outputTokens: number;
  thinkingTokens: number;
}

export interface ProcessPageInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  documentName: string;
  /** @minimum 1 */
  pageNumber: number;
  /** @minLength 32 */
  dataUrl: string;
}

export interface ProcessPageResult {
  page: ReconstructedPage;
  usage?: Usage;
}

export interface ProcessDocumentResult {
  documentName: string;
  pages: ReconstructedPage[];
  selectedFrameCount: number;
  discardedFrameCount: number;
  processingNote: string;
  usage?: Usage;
}