# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: processing
    description: Document processing operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /process-document:
    post:
      operationId: processDocument
      tags: [processing]
      summary: Analyze selected document frames
      description: Sends locally filtered document frames to Gemini for ordered text extraction and confidence scoring.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessDocumentInput"
      responses:
        "200":
          description: Reconstructed document analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessDocumentResult"
        "400":
          description: Invalid frames
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: AI usage limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: AI declined to transcribe these frames
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: AI analysis unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI temporarily busy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /process-page:
    post:
      operationId: processPage
      tags: [processing]
      summary: Rebuild one page from a retake photo
      description: Sends one fresh photo of a single page to Gemini and returns that page rebuilt, so it can replace a page that was hard to read.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessPageInput"
      responses:
        "200":
          description: The rebuilt page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessPageResult"
        "400":
          description: Invalid image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: AI declined to transcribe this photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: AI usage limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: AI analysis unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI temporarily busy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Calm, plain-language message that can be shown to the user.
        code:
          type: string
          enum: [bad_request, quota, busy, invalid_output, blocked, unavailable]
      required:
        - error
        - code
    ProcessFrame:
      type: object
      properties:
        dataUrl:
          type: string
          minLength: 32
        timestamp:
          type: number
          minimum: 0
        sharpness:
          type: number
          minimum: 0
        difference:
          type: number
          minimum: 0
      required:
        - dataUrl
        - timestamp
        - sharpness
        - difference
    ProcessDocumentInput:
      type: object
      properties:
        documentName:
          type: string
          minLength: 1
          maxLength: 160
        frames:
          type: array
          minItems: 1
          maxItems: 30
          items:
            $ref: "#/components/schemas/ProcessFrame"
      required:
        - documentName
        - frames
    Box:
      type: array
      description: "Region of an image as [ymin, xmin, ymax, xmax], each 0-1000 (fractions of the image height/width times 1000)."
      minItems: 4
      maxItems: 4
      items:
        type: integer
        minimum: 0
        maximum: 1000
    PageBlock:
      type: object
      description: One piece of a page in reading order, so the page's structure can be rebuilt.
      properties:
        type:
          type: string
          enum: [heading, paragraph, list, table, lines, caption, figure, header, footer]
        level:
          type: integer
          minimum: 1
          maximum: 3
          description: Heading level (1 is the largest). Only for headings.
        text:
          type: string
          description: Text of the block. For type "lines" each line of the page is on its own line.
        items:
          type: array
          description: List items without their bullet or number. Only for lists.
          items:
            type: string
        rows:
          type: array
          description: Table rows, each a list of cell texts. Only for tables.
          items:
            type: array
            items:
              type: string
        box:
          description: Where the block is in the best frame, or null if unknown.
          oneOf:
            - $ref: "#/components/schemas/Box"
            - type: "null"
      required:
        - type
        - text
        - box
    ReconstructedPage:
      type: object
      properties:
        pageNumber:
          type: integer
          minimum: 1
        title:
          type: string
        text:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        needsReview:
          type: boolean
        reviewReason:
          type: ["string", "null"]
        sourceFrameIndices:
          type: array
          description: Zero-based indices (into the request frames) of every frame that shows this page.
          minItems: 1
          items:
            type: integer
            minimum: 0
        bestFrameIndex:
          type: integer
          minimum: 0
          description: Zero-based index of the frame that shows this page most clearly. Always one of sourceFrameIndices.
        pageBox:
          description: Where the paper page is inside the best frame, or null if unknown.
          oneOf:
            - $ref: "#/components/schemas/Box"
            - type: "null"
        blocks:
          type: array
          description: The page's content in reading order. The same text as `text`, with structure and position.
          items:
            $ref: "#/components/schemas/PageBlock"
        modelUsed:
          type: string
          description: Gemini model that produced this page.
      required:
        - pageNumber
        - title
        - text
        - confidence
        - needsReview
        - reviewReason
        - sourceFrameIndices
        - bestFrameIndex
        - pageBox
        - blocks
    Usage:
      type: object
      description: What this request used, for working out running costs.
      properties:
        model:
          type: string
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        thinkingTokens:
          type: integer
      required:
        - model
        - inputTokens
        - outputTokens
        - thinkingTokens
    ProcessPageInput:
      type: object
      properties:
        documentName:
          type: string
          minLength: 1
          maxLength: 160
        pageNumber:
          type: integer
          minimum: 1
        dataUrl:
          type: string
          minLength: 32
      required:
        - documentName
        - pageNumber
        - dataUrl
    ProcessPageResult:
      type: object
      properties:
        page:
          $ref: "#/components/schemas/ReconstructedPage"
        usage:
          $ref: "#/components/schemas/Usage"
      required:
        - page
    ProcessDocumentResult:
      type: object
      properties:
        documentName:
          type: string
        pages:
          type: array
          items:
            $ref: "#/components/schemas/ReconstructedPage"
        selectedFrameCount:
          type: integer
        discardedFrameCount:
          type: integer
        processingNote:
          type: string
        usage:
          $ref: "#/components/schemas/Usage"
      required:
        - documentName
        - pages
        - selectedFrameCount
        - discardedFrameCount
        - processingNote