# lib/api-zod/src/generated/types/processPageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProcessPageInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  documentName: string;
  /** @minimum 1 */
  pageNumber: number;
  /** @minLength 32 */
  dataUrl: string;
}