# lib/api-zod/src/generated/types/processPageResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReconstructedPage } from './reconstructedPage';
import type { Usage } from './usage';

export interface ProcessPageResult {
  page: ReconstructedPage;
  usage?: Usage;
}