# lib/api-zod/src/generated/types/usage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * What this request used, for working out running costs.
 */
export interface Usage {
  model: string;
  inputTokens: number;
  outputTokens: number;
  thinkingTokens: number;
}